+++ gaclen/src/graphics.rs
@@ -4,20 +4,19 @@
 //! 
 //! The graphical workflow is extensive, please refer to [examples](https://github.com/Griffone/gaclen/tree/master/examples) for help.
 
+pub mod buffer;
 pub mod context;
 pub mod device;
 pub mod frame;
+pub mod image;
 pub mod pass;
 pub mod swapchain;
 
 /// used for hardware acceleration.
 pub use vulkano;
 pub use vulkano::impl_vertex;
-pub use vulkano::buffer::BufferUsage;
 pub use vulkano::format::Format as PixelFormat;
-pub use vulkano::image::Dimensions;
 pub use vulkano::instance::Version;
-pub use vulkano::sampler::{Filter as TextureFilter, SamplerAddressMode, MipmapMode};
 
 const REQUIRED_VULKAN_VERSION: Version = Version { major: 1, minor: 0, patch: 0 };
 const ENGINE_NAME: &str = "gaclen";